Fix error in simulation file, plus set a seed
It turns out our simulation was trying to search
for an output directory which had not been created!
This commit addesses that issue by adding a check
for an "output" directory, and creating it
if it does not exist.

In addition, the simulation requires the generation
of a random initial state psi_0. To ensure students
obtain the same output, the numpy's random.seed()
function is set to a particular value, thereby fixing
the random part of the simulation.