Reset session 1 files
There are a couple of issues in the session one files
which were unknown at the time they were uploaded to the
site (commit ba940577effd7d). This commit removes the .zip
file from the repository, as well as the download link
from the website.

Once the issues are addressed, a new zip file will be uploaded,
and the link re-created.