Update Python Visualization lesson
Another run-through showed that there was still
too much content. This commit cuts down on that
by getting rid of one plot example (histogram),
and also pre-filling a bit more of the pandas/seaborn
commands. (Understanding those commands is not essential
to the point of this lesson, and so, there's no reason
to make learners type them in.)